product sucesses

+++ database/migrations/2017_05_03_165853_create_orders_details_table.php
@@ -16,7 +16,7 @@ class CreateOrdersDetailsTable extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->string('order_id');
             $table->string('user_id');
-            $table->integer('total_price');
+            $table->double('total_price');
             $table->timestamps();
             $table->softDeletes();
             $table->foreign('order_id')->references('order_id')->on('orders');
